requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf_spectacular_extras==0.1.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prometheus_client==0.22.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
six==1.17.0
sqlparse==0.5.3
swagger_ui_bundle==1.1.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
